+++ src/Factorial.Messages/Events/FactorialCalculated.cs
@@ -3,6 +3,6 @@ namespace Factorial.Messages.Events
     public class FactorialCalculated : IEvent
     {
         public int n { get; set; }
-        public int Result { get; set; }
+        public ulong Result { get; set; }
     }
 }
